Refactor: Move community posts logic to community_posts.py
- Created a new file `community_posts.py` under the `routes` directory.
- Moved all community posts-related endpoints (`/community-posts`, `/post-community-post`) from `main.py` to `community_posts.py`.
- Used `APIRouter` for modular route handling.
- Updated `main.py` to include the `community_posts_router` with the prefix `/community-posts`.
- Improved code organization and maintainability by separating route logic.
